Potential fix for crashes when using many CPUs
There is a bug that prevents the usage of high multigrid levels and
many CPUs, as parosol would crash.
One cause for this were successive -1 entries in the list, which
would lead to a situation where -1 would stay in the list.
This fix iterates the list from the back and prevents such
situations.
Please test this change thoroughly!